[refactor] 모듈명 수정
