+++ movie/Dockerfile
@@ -4,13 +4,13 @@ WORKDIR /app
 
 COPY . .
 
-RUN gradle :movie-api:bootJar --no-daemon
+RUN gradle clean :api:bootJar --no-daemon
 
 FROM openjdk:21-slim
 
 WORKDIR /app
 
-COPY --from=build /app/movie-api/build/libs/*.jar app.jar
+COPY --from=build /app/api/build/libs/*.jar app.jar
 
 EXPOSE 8080
 ENTRYPOINT ["java", "-jar", "app.jar"]
