+++ movie/api/src/test/kotlin/com/example/movie/MovieApiApplicationTests.kt
@@ -1,0 +1,13 @@
+package com.example.movie
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class MovieApiApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+
+}
