+++ movie/core/src/main/kotlin/com/example/movie/domain/common/model/BaseModel.kt
@@ -1,0 +1,10 @@
+package com.example.movie.domain.common.model
+
+import java.time.LocalDateTime
+
+interface BaseModel {
+    val createdBy: String
+    val createdAt: LocalDateTime
+    val updatedBy: String
+    val updatedAt: LocalDateTime
+}
