+++ movie/core/src/main/kotlin/com/example/movie/domain/movie/model/Genre.kt
@@ -1,0 +1,14 @@
+package com.example.movie.domain.movie.model
+
+import com.example.movie.domain.common.model.BaseModel
+import java.time.LocalDateTime
+
+class Genre(
+    val id: Long,
+    val name: String,
+    val description: String,
+    override val createdBy: String,
+    override val createdAt: LocalDateTime,
+    override val updatedBy: String,
+    override val updatedAt: LocalDateTime
+) : BaseModel
