+++ movie/core/src/main/kotlin/com/example/movie/domain/movie/repository/GenreRepository.kt
@@ -1,0 +1,9 @@
+package com.example.movie.domain.movie.repository
+
+import com.example.movie.domain.movie.model.Genre
+
+interface GenreRepository {
+    fun findById(id: Long): Genre?
+    fun findByName(name: String): Genre?
+    fun findAll(): List<Genre>
+}
