+++ movie/docker-compose.yml
@@ -1,23 +1,37 @@
 services:
   movie-mysql:
     image: mysql:8.0
+    container_name: movie-mysql
     ports:
       - "3306:3306"
     environment:
       - MYSQL_DATABASE=movie_db
       - MYSQL_ROOT_PASSWORD=root
+    command:
+      - --character-set-server=utf8mb4
+      - --collation-server=utf8mb4_unicode_ci
+    restart: always
     volumes:
       - movie_data:/var/lib/mysql
+    healthcheck:
+      test: [ "CMD", "mysqladmin", "ping", "-h", "localhost", "-u", "root", "-proot" ]
+      interval: 10s
+      timeout: 5s
+      retries: 5
 
   api-server:
     build: .
+    container_name: movie-api
     image: movie-api-server:1.0.0
     environment:
+      - SPRING_PROFILES_ACTIVE=prod
       - DB_URL=movie-mysql
     ports:
       - "8080:8080"
     depends_on:
-      - movie-mysql
+      movie-mysql:
+        condition: service_healthy
+    restart: on-failure
 
 volumes:
   movie_data:
