--- movie/movie-api/build.gradle
@@ -1,28 +1,0 @@
-plugins {
-    id 'org.jetbrains.kotlin.jvm' version '1.9.25'
-    id 'org.jetbrains.kotlin.plugin.spring' version '1.9.25'
-    id 'org.springframework.boot' version '3.4.1'
-    id 'io.spring.dependency-management' version '1.1.7'
-}
-
-dependencies {
-    implementation project(':movie-core')
-    implementation project(':movie-infra')
-
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'com.fasterxml.jackson.module:jackson-module-kotlin'
-    implementation 'org.jetbrains.kotlin:kotlin-reflect'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.jetbrains.kotlin:kotlin-test-junit5'
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-}
-
-tasks.register("prepareKotlinBuildScriptModel"){}
-
-bootJar {
-    enabled = true
-}
-
-jar {
-    enabled = false
-}
