--- movie/movie-api/src/main/resources/application.yml
@@ -1,22 +1,0 @@
-spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://${DB_URL:localhost}:3306/movie_db
-    username: root
-    password: root
-
-  jpa:
-    hibernate:
-      ddl-auto: update
-    properties:
-      hibernate:
-        format_sql: true
-    show-sql: true
-
-
-server:
-  port: 8080
-logging:
-  level:
-    com.example.movieapi: debug
-    com.example.movieinfra: debug
