--- movie/movie-api/src/test/kotlin/com/example/movieapi/MovieApiApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.example.movieapi
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class MovieApiApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
