--- movie/movie-core/build.gradle
@@ -1,11 +1,0 @@
-plugins {
-    id 'org.jetbrains.kotlin.jvm' version '1.9.25'
-}
-
-dependencies {
-    testImplementation 'org.jetbrains.kotlin:kotlin-test-junit5'
-}
-
-jar {
-    enabled = true
-}
