--- movie/movie-core/src/main/kotlin/com/example/moviecore/domain/common/model/BaseModel.kt
@@ -1,10 +1,0 @@
-package com.example.moviecore.domain.common.model
-
-import java.time.LocalDateTime
-
-interface BaseModel {
-    val createdBy: String
-    val createdAt: LocalDateTime
-    val updatedBy: String
-    val updatedAt: LocalDateTime
-}
