--- movie/movie-core/src/main/kotlin/com/example/moviecore/domain/movie/repository/GenreRepository.kt
@@ -1,9 +1,0 @@
-package com.example.moviecore.domain.movie.repository
-
-import com.example.moviecore.domain.movie.model.Genre
-
-interface GenreRepository {
-    fun findById(id: Long): Genre?
-    fun findByName(name: String): Genre?
-    fun findAll(): List<Genre>
-}
