--- movie/movie-infra/build.gradle
@@ -1,31 +1,0 @@
-plugins {
-    id 'org.springframework.boot' version '3.4.1'
-    id 'io.spring.dependency-management' version '1.1.7'
-    id 'org.jetbrains.kotlin.plugin.spring' version '1.9.25'
-    id 'org.jetbrains.kotlin.plugin.jpa' version '1.9.25'
-    id 'org.jetbrains.kotlin.jvm' version '1.9.25'
-}
-
-dependencies {
-    implementation project(':movie-core')
-
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.jetbrains.kotlin:kotlin-reflect'
-    runtimeOnly 'com.mysql:mysql-connector-j'
-    testImplementation 'org.jetbrains.kotlin:kotlin-test-junit5'
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-}
-
-allOpen {
-    annotation 'jakarta.persistence.Entity'
-    annotation 'jakarta.persistence.MappedSuperclass'
-    annotation 'jakarta.persistence.Embeddable'
-}
-
-tasks.register("prepareKotlinBuildScriptModel"){}
-bootJar {
-    enabled = false
-}
-jar {
-    enabled = true
-}
