--- movie/movie-infra/src/main/kotlin/com/example/movieinfra/persistence/common/BaseEntity.kt
@@ -1,12 +1,0 @@
-package com.example.movieinfra.persistence.common
-
-import jakarta.persistence.MappedSuperclass
-import java.time.LocalDateTime
-
-@MappedSuperclass
-abstract class BaseEntity (
-    val createdBy: String,
-    val createdAt: LocalDateTime,
-    val updatedBy: String,
-    val updatedAt: LocalDateTime
-)
