+++ movie/settings.gradle
@@ -1,4 +1,4 @@
 rootProject.name = 'movie'
-include 'movie-api'
-include 'movie-core'
-include 'movie-infra'
+include 'api'
+include 'core'
+include 'infra'
